chore(cubesql): Rewrite engine first steps -- simple aggregate rewrite works
